Removed lots of abuse of pragmas to silence warnings. Should be usingthe language to do that instead for example static_cast. Also homogenised as much as possible the use of sizes and indexing of containers to use size_t in the same spirit as the STL, this in itself removed alot of warnings. Lots of duplication between defines.h and macros.h, now consolidated to one file macros.h. Removed config.mgw.h, config.msvc.h & config.bcc.h, long overdue cleanout. Added some missing License text and Copyrights from some source files.
git-svn-id: https://svn.code.sf.net/p/xlw/code/branches/xlw4/xlw4@692 948fe115-5b3c-0410-8233-88faca9dced8

+++ xlw/src/XlOpenClose.cpp
@@ -55,7 +55,7 @@ extern "C"
 			oldStreamBuf = std::cerr.rdbuf(&debuggerStreamBuf);
 			std::cerr << XLW__HERE__ << "std::cerr redirected to MSVC debugger" << std::endl;
 
-			dwRet = VirtualQuery (((LPCVOID) (&oldStreamBuf)), &theInfo,((DWORD) (sizeof (MEMORY_BASIC_INFORMATION))));
+			dwRet = static_cast<DWORD>(VirtualQuery (((LPCVOID) (&oldStreamBuf)), &theInfo,(static_cast<DWORD> (sizeof (MEMORY_BASIC_INFORMATION)))));
 			if (dwRet)
 			{
 				theHandle = ((HMODULE) (theInfo.AllocationBase));
